version 2.1.0 Modified exercise

+++ src/main/java/com/pichincha/backend/test/dto/TransactionResponseDto.java
@@ -16,6 +16,6 @@ import lombok.experimental.FieldDefaults;
 public class TransactionResponseDto {
 
   String type;
-  double amount;
+  int amount;
 
 }
